Fixed TTuple compilation error, and a crash in Delegates.
Also added "do { ... } while (false)" to all the SafeDelete, SafeReleasem SafeAddRef etc. This is to force the use of ';' when using these macros, as well as it ensures that the macro can be used in one-line if-statements without problem (Even though the standard says that this should not be used).